Compilation tests module
This CL adds a new module which is a java project that can be
used to create one-off project setups. This module will be
used to test error cases and to make sure we detect errors
and dispatch proper error messages.

I've also updated antlr to 4.5 which gets rid of the annoying
annotation processor source compatibility error:
https://github.com/antlr/antlr4/issues/487

Bug: 21735564
Change-Id: I16c04923c7d69ca40fe13139acaf87c08166ad95