Use intrinsic drawable dimensions for TextView's drawable* attributes.
Bug 23975379

drawableLeft, etc should use setCompoundDrawablesWithIntrinsicBounds
instead of setCompoundDrawables. Because a developer may set the
drawables manually, we don't want to override the bounds for
drawables that aren't data bound. Thus, the intrinsic bounds
are calculated as part of the binding adapter for just the set
drawable rather than in TextView for all drawables.

Change-Id: Ie2032f05ef80ca2b8d3b750808715852d7cfae4e