Load full SDK table
When loading methods from the SDK, we would only load
methods that have since > Application.minApi. This
way, we would keep the hash very small. On the other
hand, this algorithm break if a newer API version adds
a method that has the same signature with a subclass.

This case happened when we added setForeground to
View at API 23 but it has been in FrameLayout since 1.

Bug:24509802
Change-Id: I7ac7ca5d1aac142a9afe8cd57ab9497dfb649650