Fix error processing XML when german character is used.
Bug 24507894

The character set was not being properly detected for the XML
file when it was being loaded, so it was always using the default
character set. For Windows, this turned out to be something other
than UTF-8. Now, a library is used to detect the character set
before parsing.

Change-Id: I2ea041701de59f0e72fd18288b0f62cb2beb74e6