Fix tag check in data binding mapper
This CL fixes a bug in data binding mapper where if the View's tag does
not match one of the expected tags, it would fall through to the ext
case statemet which would result in a meaningless error.

This CL detects the case and throws an exception. This case should not
happen since it is not a supported case but happened here due to another
bug

Bug: 24745000
Change-Id: Ic5c6d4ce2da6ad043fb6fe56cd0847c98ea2472d