Override layout file metadata from module
If a layout file is inherited from a module but also exists
in the app, w~e force the app version to the metadata of the
module version. This means forcing its created class location
to be the same as the module. This usually happens when gradle
or aapt generates some layout during app compilation.

Bug: 25369165
Change-Id: I5d2002ac04d16cfe9935fe5580548344b19b4aca
(cherry picked from commit 28e7064d455e2ef9da31c817dfc05ec7405c60df)