Jack Support
This CL introduces a minor change to detact .bin files from a build
directory instead of jars because Jack does not support them.
This directory is setup by the android gradle plugin when jack is
used and passed as an argument to the processor.

I've also introduced CUSTOM_JACK environment variable which will
force data binding tests to use jack complation when set.

Note that as of this CL, android gradle plugin still has limited Jack
support so not all applications will work well. (see: b/29994342,
https://code.google.com/p/android/issues/detail?id=214851)

Bug: 29872090
Change-Id: I7e0bbf0c21e3d681ca7bd277f2a5140436184e29