Fix grouping for cast and ternary expressions.
Bug 30567255

Group and cast operations, when used with non-dynamic
expressions are not extracted to local variables. This
means that if the result is an reference and a method
or field is accessed on it, it won't be properly grouped.

For example:

 (Constants.alwaysTrue ? "a" : "ab").length()

had no grouping for the ternary, so resulted in:

 Constants.alwaysTrue ? "a" : "ab".length()

This adds the missing grouping expression.

Change-Id: I53a6da30885fa453f8943f54892e0a21384d0a9c