Ship data binding artifacts in AAR
This CL changes how data binding keeps information about library projects.
Previously, we would produce a "bin" file inside the jar to be able to
carry over information from a library into the app. Now we will use the
aar file instead and gradle will merge them into a folder.

We still keep the old code which checks the classpath jars for backward
compatibility.

This CL also adds new tests under an anchor task called
"testBackwardCompatibility" which compiles a library project with an
older android gradle plugin version and then compiles and tests an
app that depends on the library. The app is compile with ToT.

Bug: 31033895
Change-Id: I94c2b39391eece01c86e5c894e26a581d96c83db