Ensure that bindings are evaluated prior to their included bindings.
Bug 31085763

In the case where a layout with a binding includes another layout
with a binding, it is important to evaluate the bindings of the
containing before the included. This ensures that if the containing
binding has a value set, the value is set on the included binding
before evaluation.

This is part 1 of a 2-part change. This includes only the
ViewDataBinding change. The remainder will require the version
to be updated.

Change-Id: I14f57837ebc69922663dc3b806d65b4fd4a86925