Ensure that bindings are evaluated prior to their included bindings.
Bug 31085763

In the case where a layout with a binding includes another layout
with a binding, it is important to evaluate the bindings of the
containing before the included. This ensures that if the containing
binding has a value set, the value is set on the included binding
before evaluation.

This is part 2 of a 2-part change. This includes the aspects
that depend on the library and the tests.

Change-Id: I6b16c6c3f5961958259b0ca707ec2373d073658d