Check folder names before copying
This CL fixes a bug in XML parser where when a new layout folder is added,
incremental compilation would override the processed XML file with the whole
folder (because folder was listed after the xml file).
We should not be copying layout folders in any case so this CL checks
folder names before copying.

Bug: https://code.google.com/p/android/issues/detail\?id\=226396
Test: In gradle project
Change-Id: I2ce376db1948997f2e6dd2f323be147ae9445308