Simplify semantics of SetterStore.getSetterCall
Move a special behavior for subclasses of
android.databinding.ViewDataBinding from the
SetterStore.getSetterCall method to the Binding.resolveSetterCall
method. The SetterStore.getSetterCall now always returns null if the
setter is not found.

Add a reference to an ObservableInt variable in an included layout to
TestApp to verify that the CL fixes the compilation failure reported in
the bug. 

Test: ran TestApp tests
Bug: https://code.google.com/p/android/issues/detail?id=235985

Change-Id: I6ef904faa7a1b18ab4f36cd2764bb3ab8898484b