Fix thread local memory leak
Keeping the marshallar in a thread local instance creates a
memory leak in gradle since gradle keep the thread around
and re-uses and we end up creating a new marshallar for
each new thread we use. The previous change to make it thread
local was done to fix b/35345953. In this CL, i'm moving it
back to static variable but making the toXML fromXML calls
thread safe.

Bug: 37458557
Test: The case repoted in the bug works fine now.
Change-Id: Id9bb1ba256daa48d69683174788d091a404ef7b2