Add @Generated annotation if it is available
People using error prone are having problems with data binding since
it does not generate perfect code. A good way to exclude it is using
@Generated annotation but we can only use it if app has the necessary
annotation (android.jar does not have @Generated).
This CL checks classpath to decide whether to add it or not.
Also changed test app to have it but other apps don't.

Bug: 37136286
Test: existing tests pass
Change-Id: Ie72518bc690fb98a859ac30f0940c1af18fdba6a