Fix android:selectedItemPosition.
Bug 64981992

This CL makes a multi-attribute BindingAdapter for selectedItemPosition
and adapter. This allows the selected position to work after the adapter
has been set.

This CL also adds the ability to use android:selection as well as
android:selectedItemPosition. Many developers are confused because
the accessor methods don't match. They are getSelectedItemPosition()
and setSelection().

Updates the prebuilts to 1.3.3 for the next release.

Test: coming in following CL after 1.3.3 is committed

Change-Id: I7e7d0d6951241c7c14f8a846125f4044d3cbe5f4