Keep generated binding codes in libraries
This CL changes data binding to not strip binding code that is
generated for libraries.

This will allow data binding to be more incremental and only generate
binding code for the current app/module.

Generated DataBindingMapper and BR classes are still stripped as they
require final R class and/or final list of BR classes.

Bug: 68392907
Test: existing tests pass
Change-Id: I39c8c25b90ec58a8b452e34310b57acc4b77c1d1