Generate data binding docs via doclava, independently
This CL modifies data binding runtime build files to include doclava
so that we can generate online docs through it.
Since base library is part of tools build, we also generate its docs
here.

In a followup CL, i'll try to autmate this w/ bazel but right not
doclava does not have a bazel build so it is not trivial.

Bug: 69160652
Test: local
Change-Id: I8ce08191a0c1567e184919484453deec2c68ead1