Fix kotlin.Unit method references in binding layouts.
When there is a binding adapter that receives a kotlin
function which does not return, Kotlin converts it to
Function0<kotlin.Unit>. But if there is a kotlin class
with a method that does not return value, kotlin converts
it to void.
This becomes a problem for data binding because it
cannot match the two when using function references.

We cannot simply convert kotlin.Unit to void because
we still generate code in java so we have to return
a value in the implementation.

Instead, this CL changes function matcher to accept
the two as equals and also changes the code generators
to handle the case manually.

Also added a kotlin test app (finally).

Bug: 70915745
Bug: 78662035
Test: KotlinTestApp
Change-Id: I019ee7eb1dd635b12efd7725ccc22f0c63dc2d72