Detect common null check ternary ops
This CL adds logic to the Ternary expression to avoid unboxing
if the predicate is a null check.

e.g. if it is a == null ? b : a, we don't need to unbox  for the
false statement because it will not be executed if a is null.

This is not a complete nullability detection code, it only covers the
common cases.

Bug: 37127560
Test: SafeUnboxingTest
Change-Id: I7a090b7dc92f1c127cc3b63f7eab3f6c37a250a8