Enable features
This CL adds a new build type called FEATURE, which is a new
version where we generate BR classes for only runtime dependencies
(dependencies that are not inherited from base or any other feature).

This new BR file generation logic looks up the existing ids
and re-uses if the BR id matches the id in a dependency.
For any other id, we generate them with an offset (provided by gradle)
so that it cannot conflict w/ other features.

The integration test app is not a real test app right now because
features do not support instrumentation tests. It is more of a
manual testing ground until features have the capability.

Bug: 63814741
Test: in the other CL

Change-Id: I6f3feda908b3925a29ff76a29cd3659cab3f8f8d