Remove dynamic util.
This class was only necessary when we had 2 different release cycles.
We don't need it anymore, even for backwards compatibility.

if developer is in v1, the library's code is regenerated so it won't
use DynamicUtil.

If developer is in v2, it was broken for libraries so there is no
way that a library w/ dynamic util reference exist.

Bug: 73020925
Test: TestApp/IndependentLibraryTest
Change-Id: Id43ca24a096de16c89c0913ee309ba29accccdcb