Integration tests w/ AndroidX
This CL adds integration tests w/ AndroidX and fixes some androidX
issues that i've discovered during the move.

Old tests are copied to integration-tests-support so that we still
run them.

I tried to keep the tests as similar as possible but sometimes needed
to move classes (e.g. for cases where we have package private access).
I've changed the applicationId of TestApp to use a different package
so that it wo't conflict w/ the other one while testing on the same
device. For other simpler tests, we were not using android.databinding
so i just changed their package names not to conflict w/ their twins.

I've also removed App With Spaces test which does not work w/ gradle5
anymore and we were not running it anyways.

Bug: 77166878
Test: new andoridX tests
Change-Id: I86f68fd6bee2ca849c6472c0411234bb3a08b132