Fix kotlin test flake
tests set values on the binding and then execute
pending binginds. Looks like this might sometimes
flake because variables in the binding classes
are not volatile so setting them on a background
thread may possibly delay their update.

Was able to reproduce it by running the test 1000 times.
After this change, it did run w/o a flake for 1000 times.

Bug: 77902456
Test: KotlinTestApp
Change-Id: I4ebf72b198a071f382af29c9481c5f012762095f