Null check class analyzer calls
This is a quick fix to null check class anayzer
results to avoid NPE while traversing adapters.

It is tricky to test since this only happens with
a dependency that was both compiled with android.support
and pre 3.2 version. This is why I've added LibCompiledWith3.1
to the list of integration tests which is compiled manually
and the final aar is put under MultiModuleTestApp/app/libs.

Bug: 116361870
Test: MultiModuleTestApp
Change-Id: Ibaf6174980de08d50b0ef1b9679db6cf99e1121a