Fix two way binding for field access expressions
This CL fixes a bug in FIeldAccessExpr's two way resolution logic.
FieldAccessExpr is the last stop in two way resolution as it is
either targeted to a view (check for two way) or not (nothing to do).
But this breaks in cases where it might have nested access levels.
e.g. textView.text.length has a nested FieldAccessExpr which is
two way (textView.text) but the target for the parent is textView.text
which is a FieldAccessExpression.
Previous logic was stopping resolution when it is not ViewAccess but
it should've continued to call the parent to check nested items.

Bug: 116536168
Test: TwoWayBetweenViewsTest

Change-Id: I74e9c62ee21fc06f6128c6baaebde21496c495b9