Don't generate V1Compat mapper for libs
This CL fixes a bug where if a module depends on a V1 library,
we would generate the V1 mapper for it which does not compile
because R files for V1 are not final at that stage.
We should actually not generate it and instead delegate the
application to do it.
We still generate the stub binding for those so that the code
can compile but we'll strip them as usual.

Bug: 117666264
Test: MultiModuleTestApp with LibCompiledWithV1 dependency
Change-Id: I9cc6b04573776141b021d0d516f10fc0213e8270