Avoid processor generated code in gradle generated code
This CL works around an issue where JAVAC prints a lot of
undefined element errors when annotation processor fails for
an unrelated reason. Each DataBinding base class that is generated
in gradle has 5 accesses to DataBindingComponent class which
means user does not see any real errors if they have just 20
data bound layouts.

This CL changes Gradle generated code to accept Object instead
of DataBindingComponent to workaround this issue.
Even though it is ugly, these codes are mosltly called by the
generated code so that does not matter.

For developers who call these manually, they also usually call
the method without the DataBinding component so that is unchanged.

This only creates an inconvenience for people who call the method
with a DataBindingComponent but I think it is a fair inconvenience
to that use case to make the life of all other developers better.

I've also deprecated these methods and generated replacement
in the javadocs so that we can guide developers to the right
direction.

Not proud of this change but this is the only way to fix this
problem in a backwards compatible way without changing javac.

Bug: 116541301
Test: existing tests pass

Change-Id: Ibb3c0f865560d9d919287d5987caf84fe6d60cf7