Better support for generics
This CL fixes a bug where we are unable to declare local variables
with the correct type if the variable has wildcard generics.
This was because we use ModelClass's toCode method, which is not
really desired unless we are trying to access that ModelClass in
code (rather than declaring a variable with its type).

This CL changes code generater to use TypeName instead when declaring
fields.

Bug: 119960651
Bug: 116821433
Test: GenericTest in KotlinTestApp
Change-Id: I7075bd533ea63570be0ced3b80a1ea03dfd1e56b